+++ src/cli/mod.rs
@@ -3,6 +3,7 @@
 pub use crate::config::Config;
 
 pub mod exec;
+pub mod fetch;
 pub mod init;
 pub mod max;
 pub mod test;